backend/requirements.txt
fastapi>=0.115.6
uvicorn[standard]>=0.32.1
sqlalchemy>=2.0.40
python-jose[cryptography]>=3.3.0
bcrypt>=4.2.1
python-multipart>=0.0.20
python-dotenv>=1.0.1
